move file to observability folder
