+++ assets/src/components/cd/globalSettings/observability/ObservabilitySettings.tsx
@@ -13,12 +13,12 @@ import { useUpdateDeploymentSettingsMutation } from 'generated/graphql'
 import { GqlError } from 'components/utils/Alert'
 import { useUpdateState } from 'components/hooks/useUpdateState'
 
-import { StepH } from '../ModalAlt'
+import { StepH } from '../../ModalAlt'
 
 import {
   getGlobalSettingsBreadcrumbs,
   useGlobalSettingsContext,
-} from './GlobalSettings'
+} from '../GlobalSettings'
 
 function HttpForm({ name, connection, setConnection }) {
   const theme = useTheme()
